--- payment-api/build.gradle
@@ -1,3 +1,0 @@
-dependencies {
-    implementation project(':model')
-}
