+++ shop-api/src/main/java/org/healthplus/shop/application/service/MenuModificationService.java
@@ -3,7 +3,7 @@ package org.healthplus.shop.application.service;
 import lombok.RequiredArgsConstructor;
 import org.healthplus.shop.domain.entity.Menu;
 import org.healthplus.shop.domain.exception.MenuNotFoundException;
-import org.healthplus.shop.infrastructure.ShopRepository;
+import org.healthplus.shop.domain.repository.ShopRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
