+++ shop-api/src/main/java/org/healthplus/shop/application/service/ShopRegistrationService.java
@@ -2,7 +2,7 @@ package org.healthplus.shop.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.healthplus.shop.domain.entity.Shop;
-import org.healthplus.shop.infrastructure.ShopRepository;
+import org.healthplus.shop.domain.repository.ShopRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
