+++ shop-api/src/main/java/org/healthplus/shop/application/service/ShopRemovalService.java
@@ -3,7 +3,7 @@ package org.healthplus.shop.application.service;
 import lombok.RequiredArgsConstructor;
 import org.healthplus.shop.domain.exception.ShopNotFoundException;
 import org.healthplus.shop.domain.entity.Shop;
-import org.healthplus.shop.infrastructure.ShopRepository;
+import org.healthplus.shop.domain.repository.ShopRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
