+++ shop-api/src/main/java/org/healthplus/shop/application/service/VendorRemovalService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.healthplus.shop.domain.exception.ShopNotFoundException;
 import org.healthplus.shop.domain.exception.VendorNotFoundException;
 import org.healthplus.shop.domain.entity.Shop;
-import org.healthplus.shop.infrastructure.ShopRepository;
-import org.healthplus.shop.infrastructure.VendorRepository;
+import org.healthplus.shop.domain.repository.ShopRepository;
+import org.healthplus.shop.domain.repository.VendorRepository;
 import org.healthplus.shop.domain.entity.Vendor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
