+++ shop-api/src/main/java/org/healthplus/shop/infrastructure/VendorRepositoryAdapter.java
@@ -2,6 +2,7 @@ package org.healthplus.shop.infrastructure;
 
 import lombok.RequiredArgsConstructor;
 import org.healthplus.shop.domain.entity.Vendor;
+import org.healthplus.shop.domain.repository.VendorRepository;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
